stop ckpt tarballs from growing with each ckpt/restore
This was the original bug report:

I discovered that image files grow with each successive checkpoint.  What
seems to be going on is that podman untars the image into a dir named
/var/lib/containers/storage/overlay-containers/<<<CONTAINERID>>>/userdata/checkpoint
, and restores from there.  But it doesn't delete this directory after it's
done; a successive checkpoint will just add to it.  The solution is to
manually identify/delete this dir after the checkpointing is finished.  The
tricky bit is that you'll have to figure out the container's ID -- some
googling reveals it's possible to get with podman inspect --format="{{.Id}}"
CONTAINER_NAME

Solution: I commented out the "--keep" option in podman restore.  It would
still be good to keep the restore log that that generates, but not the
checkpoint files.